Move mapping issues, licenses, notifications, campaigns, organisations, users and system related functions for backend api
