+++ server/api/system/statistics.py
@@ -1,26 +1,26 @@
 from flask_restful import Resource, current_app
-from server.services.settings_service import SettingsService
+from server.services.stats_service import StatsService
 
 
-class LanguagesAPI(Resource):
+class SystemStatisticsAPI(Resource):
     def get(self):
         """
-        Gets all supported languages
+        Get HomePage Stats
         ---
         tags:
-          - settings
+          - system
         produces:
           - application/json
         responses:
             200:
-                description: Supported Languages
+                description: Project stats
             500:
                 description: Internal Server Error
         """
         try:
-            languages = SettingsService.get_settings()
-            return languages.to_primitive(), 200
+            stats = StatsService.get_homepage_stats()
+            return stats.to_primitive(), 200
         except Exception as e:
-            error_msg = f"Languages GET - unhandled error: {str(e)}"
+            error_msg = f"Unhandled error: {str(e)}"
             current_app.logger.critical(error_msg)
             return {"error": error_msg}, 500
