+++ server/api/users/statistics.py
@@ -1,0 +1,37 @@
+from flask_restful import Resource, current_app
+from server.services.users.user_service import UserService, NotFound
+
+
+class UsersStatisticsAPI(Resource):
+    def get(self, username):
+        """
+        Get detailed stats about user
+        ---
+        tags:
+          - users
+        produces:
+          - application/json
+        parameters:
+            - name: username
+              in: path
+              description: The users username
+              required: true
+              type: string
+              default: Thinkwhere
+        responses:
+            200:
+                description: User found
+            404:
+                description: User not found
+            500:
+                description: Internal Server Error
+        """
+        try:
+            stats_dto = UserService.get_detailed_stats(username)
+            return stats_dto.to_primitive(), 200
+        except NotFound:
+            return {"Error": "User not found"}, 404
+        except Exception as e:
+            error_msg = f"User GET - unhandled error: {str(e)}"
+            current_app.logger.critical(error_msg)
+            return {"error": error_msg}, 500
